requirements.txt
# Core RAG Requirements (Minimal)
python-dotenv
pandas
numpy
faiss-cpu
openai
ollama
tiktoken
pytest

